+++ pos_accounting/src/pool/tests/mod.rs
@@ -13,9 +13,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common::primitives::H256;
+use std::collections::BTreeMap;
 
-use crate::{DelegationId, PoolId};
+use common::primitives::{Amount, H256};
+use crypto::{
+    key::{KeyKind, PrivateKey, PublicKey},
+    random::{CryptoRng, Rng},
+};
+
+use crate::{
+    storage::in_memory::InMemoryPoSAccounting, DelegationData, DelegationId, PoolData, PoolId,
+};
 
 mod delta_tests;
 mod operations_tests;
@@ -27,4 +35,50 @@ fn new_pool_id(v: u64) -> PoolId {
 
 fn new_delegation_id(v: u64) -> DelegationId {
     DelegationId::new(H256::from_low_u64_be(v))
+}
+
+fn create_storage_with_pool(
+    rng: &mut (impl Rng + CryptoRng),
+    pledged_amount: Amount,
+) -> (PoolId, PublicKey, InMemoryPoSAccounting) {
+    let pool_id = new_pool_id(rng.next_u64());
+    let (_, pub_key) = PrivateKey::new_from_rng(rng, KeyKind::RistrettoSchnorr);
+
+    let storage = InMemoryPoSAccounting::from_values(
+        BTreeMap::from([(pool_id, PoolData::new(pub_key.clone(), pledged_amount))]),
+        BTreeMap::from([(pool_id, pledged_amount)]),
+        BTreeMap::new(),
+        BTreeMap::new(),
+        BTreeMap::new(),
+    );
+    (pool_id, pub_key, storage)
+}
+
+fn create_storage_with_pool_and_delegation(
+    rng: &mut (impl Rng + CryptoRng),
+    pledged_amount: Amount,
+    delegated_amount: Amount,
+) -> (
+    PoolId,
+    PublicKey,
+    DelegationId,
+    PublicKey,
+    InMemoryPoSAccounting,
+) {
+    let pool_id = new_pool_id(rng.next_u64());
+    let delegation_id = new_delegation_id(rng.next_u64());
+    let (_, pub_key_pool) = PrivateKey::new_from_rng(rng, KeyKind::RistrettoSchnorr);
+    let (_, pub_key_del) = PrivateKey::new_from_rng(rng, KeyKind::RistrettoSchnorr);
+
+    let storage = InMemoryPoSAccounting::from_values(
+        BTreeMap::from([(pool_id, PoolData::new(pub_key_pool.clone(), pledged_amount))]),
+        BTreeMap::from([(pool_id, (pledged_amount + delegated_amount).unwrap())]),
+        BTreeMap::from([((pool_id, delegation_id), delegated_amount)]),
+        BTreeMap::from([(delegation_id, delegated_amount)]),
+        BTreeMap::from([(
+            delegation_id,
+            DelegationData::new(pool_id, pub_key_del.clone()),
+        )]),
+    );
+    (pool_id, pub_key_pool, delegation_id, pub_key_del, storage)
 }
